Project can be selected on Project Statistics page now
- It was not possible to change project selection on project statistics page
- We had to go back to sessions page and select project on session to see it's statistics
- Break down statistics page in to component and page for reuse
