+++ src/Pixel.Automation.Web.Portal/Pages/SessionDetailsPage.razor.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Components;
-using MudBlazor;
 using Pixel.Automation.Web.Portal.Services;
 using Pixel.Automation.Web.Portal.ViewModels;
 using Pixel.Persistence.Core.Models;
